Checkpoint_3 Global fix
