+++ bin/custom/questions/resources/impex/essentialdata-email.impex
@@ -22,12 +22,3 @@ INSERT_UPDATE EmailPageTemplate; $contentCV[unique=true]; uid[unique=true];name;
 
 INSERT_UPDATE EmailPage; $contentCV[unique=true]; uid[unique=true]; name; masterTemplate(uid,$contentCV); defaultPage; approvalStatus(code)[default='approved']; fromEmail[lang=en]; fromName[lang=en]
 ;;TrainingEmail; Training Email; TrainingEmailTemplate; true;;programmersworldminsk@gmail.com; System
-
-#INSERT_UPDATE ContentSlotName; name[unique=true]; template(uid, $contentCV)[unique=true][default='TrainingEmailTemplate'];validComponentTypes(code)
-#                             ; SiteLogo;;;logo
-#                             ; TopContent;;$wideContent;
-#                             ; BottomContent;;$wideContent;
-#
-#INSERT_UPDATE ContentSlotForTemplate; $contentCV[unique=true];uid[unique=true];position[unique=true];pageTemplate(uid,$contentCV)[unique=true][default='TrainingEmailTemplate'];contentSlot(uid,$contentCV)[unique=true];allowOverwrite
-#;;SiteLogo-TrainingEmail;SiteLogo;;EmailSiteLogoSlot;true
-#;;BottomContent-TrainingEmail;BottomContent;;EmailBottomSlot;true
