+++ bin/custom/questions/resources/questions-items.xml
@@ -9,6 +9,10 @@
 <items xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:noNamespaceSchemaLocation="items.xsd">
 
+    <collectiontypes>
+        <collectiontype code="QuestionsList" elementtype="Question" generate="true" type="list"/>
+    </collectiontypes>
+
     <relations>
         <relation code="Product2Question" localized="false">
             <sourceElement type="Product" qualifier="product" cardinality="one">
@@ -19,7 +23,6 @@
             </targetElement>
         </relation>
     </relations>
-
 
     <itemtypes>
         <itemtype code="Question"
@@ -89,6 +92,10 @@
                     <defaultvalue>shubelkom@gmail.com</defaultvalue>
                     <persistence type="property"/>
                 </attribute>
+                <attribute qualifier="lastRunnedTime" type="java.util.Date">
+                    <modifiers read="true" write="true" optional="true"/>
+                    <persistence type="property"/>
+                </attribute>
             </attributes>
         </itemtype>
 
@@ -102,7 +109,7 @@
                     <modifiers read="true" write="true" unique="false" optional="true"/>
                     <persistence type="property"/>
                 </attribute>
-                <attribute qualifier="content" type="java.lang.String">
+                <attribute qualifier="questions" type="QuestionsList">
                     <modifiers read="true" write="true" unique="false" optional="true"/>
                     <persistence type="property"/>
                 </attribute>
